Update patch set 4
Patch Set 4: Verified; I would prefer that you didn't submit this

(6 inline comments)

In general this looks very good. Nice that you caught a bug or two!

I think we need to clarify who is responsible of making sure that the x->e_mbd data is either untouched or properly reset to its right state after decoding. Made some comments on that.

Also ran this and verified that it looked good. Seems to look about the same as the previous version.

Patch-set: 4
Label: Verified=+1