Update patch set 4
Patch Set 4: (4 inline comments)

A few minor nits.. 

For example you always seem to add a space before the final ;

Deb, I will leave it to you to decide when this is ready to merge but the changes will need to be ported into the new re-styled code-base. I suggest we do this sooner rather than later.

Patch-set: 4
Label: Code-Review=0
Label: Verified=0