Update patch set 1
Patch Set 1: Looks good to me, approved

This LGTM as-is. x32 is a ABI for the x68_64 processor, so if we were to distinguish it in the target tuple it'd be in the second word, not the first. If we can detect the target ABI from the toolchain, I don't see the need to make it explicit at this point.

Patch-set: 1
Reviewer: John Koleszar <1000856@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2
Label: SUBM=+1
Label: Verified=0