Update patch set 1
Patch Set 1: Looks good to me, approved

I guess this is OK. I was hoping to make the init_flags a setting in the class rather than a function argument (and then child classes can override the default parent value of 0). We can always do that later as a refactor.

Patch-set: 1
Label: Code-Review=+2
Label: Verified=0