Update patch set 1
Patch Set 1:

So given all the code changes and stuff needed to make this work, an the stress that would cause if one would want to SIMD this (and complexity added in other parts of the code), I wonder if design-wise it makes sense to alter this model to (instead of adding a new mode) instead keep all existing modes, and use this model to predict the tree probabilities instead of as an actual explicit mode? If you do that, all other changes in the code are not necessary anymore.

Patch-set: 1
Label: Code-Review=0
Label: Verified=0