Update patch set 1
Patch Set 1:

I want to make sure that we're not working around undefined behavior. If this is really a compiler bug, then it should be worked around for only that compiler (_MSC_VER == 1500 I think), but if this is a case where we're doing something undefined and the compiler has the freedom to compile it the way it did, then we should fix our code.

Any data on my earlier question about the alignment of the pointer?

Patch-set: 1