Update patch set 6
Patch Set 6:

I'll ask him. If it's really an issue, it'd be better to remove these consts than to add them -- it'd just make the prototypes harder to read with no benefit.

I love const correctness of the pointed-to *data* -- we should absolutely be changing 'foo *' to 'const foo *' where we can, but making the pointers themselves const 'foo *const' excessive to me. I mean, if these pointers should be const, why not nearly every function argument? Should we change to:

  void write_bit(coder* const bc, const int bit, const int prob);

And stylistically it's weird, becuase if you write the arguments out on each line, you probably want to write it as:

  void write_bit(coder* const bc,
                 int const    bit,
                 int const    prob)

and writing 'int const' is pretty unnatural to me.

Patch-set: 6
Label: Code-Review=0
Label: Verified=0