Update patch set 2
Patch Set 2:

Just to be clear, this changes public symbols exported in the libvpx_g.a binary. It doesn't change types, macros, filenames and static/inline functions. Those will be done, I'm working on them in a separate patch, but they are slightly less critical (since they don't cause symbol collisions), so they don't need to be done in the same patch.

Patch-set: 2