Update patch set 2
Patch Set 2:

squash, imo.. I think it may be useful to prefix these consistently, but that goes for a number of other functions too. This patch is consistent with the current transform of removing the vp9_ prefix where not necessary. I'd rather do a cleanup of the bool coder stuff and remove some of the nonsense wrapper macros with the treecoding stuff in a later commit.

Patch-set: 2
Reviewer: John Koleszar <1000856@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0
Label: Verified=0