Update patch set 1
Patch Set 1:

The segment code was mainly put in place when we only had 4x4 transforms so is one of the areas that nees heavy review.

Regarding your two questions. Until we decide what to do the EOB feature bits should go to 8 to cater for 16x16 TX yes.

The second question it looks like the constant 16 was put in in case the EOB position chosen for the 8x8 was > 16 and therefore a problem for the 4x4 TX U and V. I would have thought that using your nice new clamping function get_eob(xd, segment_id, 16) would do... so if the EOB was 0 for example it would stay 0 not jump to 16 for the UV.

Patch-set: 1
Reviewer: Paul Wilkins <1000899@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0
Label: Verified=0