Update patch set 1
Patch Set 1:

The 4x4 code indeed becomes a little ... unfortunate, because of the between-block data dependencies that Debargha's B_CONTEXT_PRED mode adds. Without that, it's a lot prettier.

Anyway, the goal here is primarily to not use vp9_decode_coefs_4x4() as a proxy towards decode_coefs() except if we're a Y block. That function adds some overhead, and that is unnecessary for non-Y blocks (Y2/UV). This should - at least in theory - be somewhat of a speed boost, especially once we mark decode_coefs() to be inlined.

Anyway, this is arguably not super-pretty, as I just said. I don't really have any ideas on how to make that better without keeping a vp9_decode_coefs_4x4() function in, or adding another function like it that lacks the overhead for Y blocks and can then be inlined also...

Patch-set: 1
Label: Verified=0