Update patch set 12
Patch Set 12:

Oops, typos in last comment, let me try that again:

Last patch should be ready to review, it isn't 100% ok but mostly works.

With experiment turned off, we go down by about -0.2% (derf), -0.3% (std-hd) to -0.6% (hd). A few clips show quite significant (>1%) drop, e.g. crew drops 2% on a few bitrates with the experiment itself turned off. This may be because of the change in scan order (2x2 32x32 SBs in 1 64x64 SB), but may also suggest some small bugs.

With experiment turned ON (compared to before this patch, not compared to experiment-turned-off), derf is slightly down (-0.1%), hd is up by about 1%, and std-hd is up by 3%.

The patch is kind of big, it includes refactoring necessary for the 2x2 32x32 SB scan ordering in order to allow for 64x64 SBs in the bitstream. If wanted, I can try to split that in a slightly more orderly fashion. Unfortunately, in the end, yes you really need these many changes...

Patch-set: 12