Update patch set 1
Patch Set 1:

I think order for it to be truly error resilient we also have to disable usage of the last mv in the find-nearest set of calls.   

Also I think a unit test to prove its error resilience is in order.    I think you should ask adrian for that, he wrote one but didn't finish..  

Basically if you set up an encode in which you code frame 0, 1 and force frame 2 to use only frame 0 as a reference.  With error resilient mode on frame 2 should decode just fine.

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0