Update patch set 6
Patch Set 6: I would prefer that you didn't submit this

I just wanted to reiterate my comments from patch set 1 -- this change may be better for your workflow right now, but is harmful for ordinary users and anyone who wants to script vpxenc:

I think in general exiting on the first error is the best thing to do and should be the default. No objection if you want to introduce an option to continue in the presence of errors -- I'd combine it with updating the mismatch check at 2354 to be fatal by default, consistent with this check, unless the option (--continue?) is given. In either case, the program should exit with an error when it finishes.

Patch-set: 6
Label: Code-Review=-1
Label: Commit-Queue=0
Label: Verified=0