Update patch set 1
Patch Set 1:

Maybe Scott can comment further, but I'm a little concerned about the horizontal blocksize parameter in this function. Since the caller already knows the blocksize, having special convolve64/32/16/8/4 functions would be easier to implement in straight assembly - and somewhat faster. (I realize that this leads to some minor code duplication in the C code. Maybe we can macro'ize the C code somewhat to make this less of a pain?)

I like the removal (for simpification purposes) of the 6- and 2-tap, we can re-add that later if there's a specific performance need. For now, this is less code clutter.

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0