Update patch set 5
Patch Set 5: I would prefer that you didn't submit this

Looks very good overall. The only problem I found is that 
the sizes of all the tiles seems to be after the frame headers. This makes it impossible to send the frame headers and the first tile before the whole frame is encoded. Do you see any problems with having the size of each tile in the beginning of the tile?

Patch-set: 5
Label: Code-Review=-1
Label: Commit-Queue=0