Update patch set 5
Patch Set 5:

You definitely want the design to be able to support sending the data for each tile as soon as it's done, you can't wait for the whole frame to be done. That's important for latency in the RTC case. Also, it'd be preferable to write the headers with each tile. At least an option to do so, but preferably the only way to do it to keep it simple. Then you can send each tile separately via RTP without creating any dependencies on a common header.

You may not even need to worry about writing the tile lengths into the bitstream. Ie, if you're tiling, we can return the tiles individually to the application, and it's responsible for keeping the lengths around (ie via RTP), and passing them back to us at decode time. If you concatenate the tiles for storing in a file (like we're doing with alt ref now), then they can be decoded serially and you can still do parallelism via frame threading.

Patch-set: 5
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0