Update patch set 1
Patch Set 1:

Yea, that is what I said.  For this clip, most of the time was spent in the idct32x32 and eob == 1 for 98% (guessing) of the time. So, it wasn't very hard to get a boost.  I suspect other clips will not show much.

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0