Update patch set 1
Patch Set 1:

I have experiments in progress that deal with other aspects of this including the default entropy values which currently make it hard to code Nearest if there are non non zero values available.

Re the removal of the rd check. I actually meant that change to go in here because it does impact the ability of the nearest mode to be used as 0,0 in compound prediction but forgot to merge it. I will do so now but leave the entropy fixes to a later patch when I can work through the implications.

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0