Update patch set 2
Patch Set 2:

There is a trade off between the chance of a more entropy efficient way of coding a mix of 0,0 and non zero vs the loss of an alternative non zero predictor to avoid coding a new vector.

Allowing Nearest to be 0 is likely to be a positive trade off as this will then be usable when there are no non zero candidates. The Near case though is typically more expensive to encode and is not often likely to be a good alternate way of coding 0,0; so the balance falls in favour of having the extra non zero candidate if one is avaliable.

Patch-set: 2
Reviewer: Paul Wilkins <1000899@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0