Update patch set 2
Patch Set 2: Looks good to me, approved

I am OK with this going in. Even if the quality on HD is down I think that the behaviour can be tuned differently to get back to where we were if that proves to be the case.

Patch-set: 2
Reviewer: Paul Wilkins <1000899@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2
Label: Commit-Queue=0
Label: Verified=0