Update patch set 3
Patch Set 3:

The bitstream sent in such a scenario isn't going to be the same one that you'd have in a webm file. The application is going to be working with individual tiles, and might not even send the lengths, if they're redundant with the encapsulation it provides. Actually it doesn't really matter much if the lengths are up front or at the end, since the application doesn't care if it's providing its own encapulsation. In general, it's best to avoid backtracking through the stream though.

Patch-set: 3