Update patch set 3
Patch Set 3:

Seems that the proposed implementation allows up to 64 column tiles. So that shouldn't be a problem.

It's true that HEVC 4k dec needs the extra memory anyway. But with 2k max tile, a 2k HEVC/VP9 dec would be able to decode HEVC upto 2k reso and VP9 without restrictions. Wouldn't that be a great benefit for VP9? Still, it would lose some compression with 4k resolution so it's a tough decision.

Patch-set: 3