Update patch set 2
Patch Set 2:

WRT the encoder: I don't think the overhead is _HUGE_. For this part, we're not filtering anywhere we weren't filtering before. Initial tests when we were prototyping this put the overhead of the scaling filter at around 15% more than a non-scaling subpel filter by hacking the existing asm. It'd be higher in cases where you could have copied instead of filtering, of course. And we have the option of adding a narrower filter if we need to reduce the complexity further. Time will tell. For motion searching, my current plan is to scale the reference buffer first, then search in the scaled buffer. For all modes, the final distortion will be taken from the common reconstruction functions, which will scale if needed.

I don't follow your second comment B, since you're describing exactly what I'm trying to do -- you ask for a filtered result given some reference and motion vector, and it'll return the (perhaps scaled)+filtered result. None of this code is concerned with the filter selection, other than setting the pointers to the appropriate bank.

Patch-set: 2
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0