Update patch set 2
Patch Set 2: (10 inline comments)

It might have been easier to add it to MACROBLOCKD, but not all of these functions get that as an argument (some take MACROBLOCK, for example) and several of these callees mutate the value in the SB case, which implies the value should probably be on the stack. Code that does this:

  tmp_foo = x->foo;
  bar(x);
  x->foo = tmp_foo;

should be avoided, so I chose to do it this way.

Patch-set: 2
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0